additions

+++ app/Actions/Jetstream/DeleteUser.php
@@ -3,6 +3,7 @@
 namespace App\Actions\Jetstream;
 
 use App\Models\User;
+use Illuminate\Validation\ValidationException;
 use Laravel\Jetstream\Contracts\DeletesUsers;
 
 class DeleteUser implements DeletesUsers
@@ -12,6 +13,12 @@ class DeleteUser implements DeletesUsers
      */
     public function delete(User $user): void
     {
+        if ($user->hasProtectedWorkOwnership()) {
+            throw ValidationException::withMessages([
+                'password' => ['Reassign your projects and created or assigned tasks before deleting this account.'],
+            ]);
+        }
+
         $user->deleteProfilePhoto();
         $user->tokens->each->delete();
         $user->delete();
